create lib and try to send struct data over tcp

+++ ur5server/ur5-server.c
@@ -164,7 +164,7 @@ int main(int argc, char **argv)
 		close(sockfd);
 
 		puts("\r\nconnection to client closed");
-		printf("average connection time: %10.4fmsec\n", (average_time/iterations)/1000);
+		printf("average connection time: %10.4fmsec\n", (float)(average_time/iterations)/1000);
 		printf("packages send: %d\n", packages_send);
 		printf("packages recieved: %d\n", packages_recieved);
 		printf("errors occured: %d\n", errors);
